Renaming of g_proximal.h to make sense
