+++ cpp/sopt/imaging_forward_backward.h
@@ -14,7 +14,7 @@
 #include "sopt/proximal.h"
 #include "sopt/relative_variation.h"
 #include "sopt/types.h"
-#include "sopt/g_proximal.h"
+#include "sopt/non_differentiable_func.h"
 
 #ifdef SOPT_MPI
 #include "sopt/mpi/communicator.h"
