+++ cpp/sopt/non_differentiable_func.h
