+++ cpp/sopt/tf_g_proximal.h
@@ -13,7 +13,7 @@
 #include "sopt/proximal.h"
 #include "sopt/relative_variation.h"
 #include "sopt/types.h"
-#include "sopt/g_proximal.h"
+#include "sopt/non_differentiable_func.h"
 
 #include "cppflow/cppflow.h"
 #include "cppflow/ops.h"
